Factor out `JsonResponse` class from `DocumentV1ApiHandler`
* Factor out other dependencies needed to make it work at a
  package-level.
* Disambiguate `JsonResponse` factory methods.
